+++ noodles-bam/src/async/reader/record.rs
@@ -113,7 +113,7 @@ async fn read_seq<R>(reader: &mut R, sequence: &mut Sequence, l_seq: usize) -> i
 where
     R: AsyncRead + Unpin,
 {
-    sequence.set_base_count(l_seq);
+    sequence.set_len(l_seq);
 
     let seq = sequence.as_mut();
     let seq_len = (l_seq + 1) / 2;