+++ noodles-bam/src/async/writer.rs
@@ -257,7 +257,7 @@ where
     let flag = u16::from(record.flags());
     writer.write_u16_le(flag).await?;
 
-    let l_seq = u32::try_from(record.sequence().base_count())
+    let l_seq = u32::try_from(record.sequence().len())
         .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_u32_le(l_seq).await?;