+++ noodles-bam/src/writer.rs
@@ -269,7 +269,7 @@ where
     let flag = u16::from(record.flags());
     writer.write_u16::<LittleEndian>(flag)?;
 
-    let l_seq = u32::try_from(record.sequence().base_count())
+    let l_seq = u32::try_from(record.sequence().len())
         .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_u32::<LittleEndian>(l_seq)?;